requirements.txt
python-docx==1.1.0
pyttsx3==2.90